chore: refactor and expose types
