+++ package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@netlify/functions",
+  "name": "@sgrove/netlify-functions",
   "main": "./dist/main.js",
   "types": "./dist/main.d.ts",
-  "version": "0.8.0",
+  "version": "0.8.0-5",
   "description": "JavaScript utilities for Netlify Functions",
   "files": [
     "dist/**/*.js",
