+++ src/function/index.ts
@@ -2,3 +2,5 @@ export { Context as HandlerContext } from './context'
 export { Event as HandlerEvent } from './event'
 export { Handler, HandlerCallback } from './handler'
 export { Response as HandlerResponse } from './response'
+export { getSecrets, withSecrets } from '../lib/secrets'
+export { NetlifySecrets } from '../lib/secrets_helper'
