+++ src/lib/secrets_helper.ts
@@ -2,7 +2,7 @@ import { Buffer } from 'buffer'
 import { request } from 'https'
 import { env } from 'process'
 
-import { HandlerEvent } from '../function'
+import { Event as HandlerEvent } from '../function/event'
 
 const services = {
   gitHub: null,
@@ -139,10 +139,12 @@ export type HandlerEventWithOneGraph = HandlerEvent & { _oneGraph: OneGraphPaylo
 
 // Note: We may want to have configurable "sets" of secrets,
 // e.g. "dev" and "prod"
-export const getSecrets = async (event: HandlerEventWithOneGraph | undefined): Promise<NetlifySecrets> => {
+export const getSecrets = async (
+  event?: HandlerEventWithOneGraph | HandlerEvent | undefined,
+): Promise<NetlifySecrets> => {
   // Allow us to get the token from event if present, else fallback to checking the env
   // eslint-disable-next-line no-underscore-dangle
-  const eventToken = event?._oneGraph?.authlifyToken
+  const eventToken = (event as HandlerEventWithOneGraph)?._oneGraph?.authlifyToken
   const secretToken = eventToken || env.ONEGRAPH_AUTHLIFY_TOKEN
 
   if (!secretToken) {
